Cleaned up jelastic maven plugin
