+++ src/main/java/com/jelastic/DeployMojo.java
@@ -1,12 +1,9 @@
 package com.jelastic;
 
 import com.jelastic.api.environment.response.NodeSSHResponses;
-import com.jelastic.model.Authentication;
-import com.jelastic.model.CreateObject;
-import com.jelastic.model.LogOut;
-import com.jelastic.model.UpLoader;
+import com.jelastic.util.PackageTypeEnum;
+import com.jelastic.util.UploadResponse;
 import org.apache.maven.plugin.MojoExecutionException;
-import org.apache.maven.plugin.MojoFailureException;
 
 /**
  * Goal which deploy artifact to Jelastic Cloud Platform
@@ -14,83 +11,37 @@ import org.apache.maven.plugin.MojoFailureException;
  * @goal deploy
  * @phase install
  */
-public class DeployMojo extends JelasticMojo {
-
-    public void execute() throws MojoExecutionException, MojoFailureException {
-        if (!isWar()) {
-            getLog().info("Skiping deploy artifact. Artifact packaging not WAR or EAR or JAR");
+public class DeployMojo extends AbstractJelasticMojo {
+    public void execute() throws MojoExecutionException {
+        if (!PackageTypeEnum.isSupported(getPackaging())) {
+            getLog().info("Skipping deploy artifact. Artifact packaging " + getPackaging() + " is not supported");
             return;
         }
-
-        Authentication authentication = authentication();
-        if (authentication.getResult() == 0) {
-            getLog().info("------------------------------------------------------------------------");
-            getLog().info("   Authentication : SUCCESS");
-            getLog().info("          Session : " + authentication.getSession());
-            //getLog().info("              Uid : " + authentication.getUid());
-            getLog().info("------------------------------------------------------------------------");
-
-            UpLoader upLoader = upload(authentication);
-            if (upLoader.getResult() == 0) {
-                getLog().info("      File UpLoad : SUCCESS");
-                getLog().info("         File URL : " + upLoader.getFile());
-                getLog().info("        File size : " + upLoader.getSize());
+        getLog().info("File Upload: Starting");
+        UploadResponse uploadResponse = upload();
+        if (uploadResponse.getResult() == 0) {
+            if (uploadResponse.getResult() == 0) {
+                getLog().info("File Upload : SUCCESS");
+                getLog().info("File URL : " + uploadResponse.getFile());
+                getLog().info("File size : " + uploadResponse.getSize());
                 getLog().info("------------------------------------------------------------------------");
 
-                CreateObject createObject = createObject(upLoader, authentication);
-                if (createObject.getResult() == 0 && createObject.getResponse().getResult() == 0) {
-                    getLog().info("File registration : SUCCESS");
-                    getLog().info("  Registration ID : " + createObject.getResponse().getObject().getId());
-                    getLog().info("     Developer ID : " + createObject.getResponse().getObject().getDeveloper());
-                    getLog().info("------------------------------------------------------------------------");
-
-                    if (isUploadOnly()) return;
-                    NodeSSHResponses deploy = deploy(authentication, upLoader);
-                    if (deploy.getResult() == 0) {
-                        getLog().info("      Deploy file : SUCCESS");
-                        getLog().info("       Deploy log :");
-                        getLog().info(deploy.getRaw());
-
-                        if (System.getProperty("jelastic-session") == null) {
-                            LogOut logOut = logOut(authentication);
-                            if (logOut.getResult() == 0) {
-                                getLog().info("           LogOut : SUCCESS");
-                            } else {
-                                getLog().info("LogOut : FAILED");
-                                getLog().error("Error : " + logOut.getError());
-                                throw new MojoExecutionException(logOut.getError());
-                            }
-                        }
-                    } else {
-                        getLog().error("          Deploy : FAILED");
-                        getLog().error("           Error : " + deploy.getError());
-
-                        throw new MojoExecutionException(deploy.getError());
-                    }
-                } else {
-                    if (createObject.getResult() != 0) {
-                        getLog().error("Create object : FAILED");
-                        getLog().error("        Error : " + createObject.getError());
-
-                        throw new MojoExecutionException(createObject.getError());
-                    } else if (createObject.getResponse().getResult() != 0) {
-                        getLog().error("Create object  : FAILED");
-                        getLog().error("Internal error : " + createObject.getResponse().getError());
-
-                        throw new MojoExecutionException(createObject.getResponse().getError());
-                    }
+                if (isUploadOnly()) {
+                    return;
                 }
-            } else {
-                getLog().error("File upload : FAILED");
-                getLog().error("      Error : " + upLoader.getError());
-
-                throw new MojoExecutionException(upLoader.getError());
+                NodeSSHResponses deploy = deploy(uploadResponse.getName(), uploadResponse.getFile());
+                if (deploy.getResult() == 0) {
+                    getLog().info("Deploy file : SUCCESS");
+                    getLog().info(deploy.getRaw());
+                } else {
+                    getLog().error("Deploy : FAILED");
+                    getLog().error("Error : " + deploy.getError());
+                }
             }
         } else {
-            getLog().error("Authentication : FAILED");
-            getLog().error("         Error : " + authentication.getError());
-
-            throw new MojoExecutionException(authentication.getError());
+            getLog().error("File upload : FAILED");
+            getLog().error("Error : " + uploadResponse.getError());
+            throw new MojoExecutionException("Upload failed : " + uploadResponse.getError());
         }
     }
 }
