--- src/main/java/com/jelastic/model/Archive.java
@@ -1,85 +1,0 @@
-package com.jelastic.model;
-
-import org.codehaus.jackson.annotate.JsonIgnoreProperties;
-
-/**
- * Created by doba on 6/17/15.
- */
-@JsonIgnoreProperties(ignoreUnknown = true)
-public class Archive {
-    private int id;
-    private int developer;
-    private String link;
-    private String name;
-    private String archive;
-    private String comment;
-    private long uploadDate;
-    private long size;
-
-    public Archive() {
-    }
-
-    public int getId() {
-        return id;
-    }
-
-    public void setId(int id) {
-        this.id = id;
-    }
-
-    public int getDeveloper() {
-        return developer;
-    }
-
-    public void setDeveloper(int developer) {
-        this.developer = developer;
-    }
-
-    public String getLink() {
-        return link;
-    }
-
-    public void setLink(String link) {
-        this.link = link;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getArchive() {
-        return archive;
-    }
-
-    public void setArchive(String archive) {
-        this.archive = archive;
-    }
-
-    public String getComment() {
-        return comment;
-    }
-
-    public void setComment(String comment) {
-        this.comment = comment;
-    }
-
-    public long getUploadDate() {
-        return uploadDate;
-    }
-
-    public void setUploadDate(long uploadDate) {
-        this.uploadDate = uploadDate;
-    }
-
-    public long getSize() {
-        return size;
-    }
-
-    public void setSize(long size) {
-        this.size = size;
-    }
-}
