--- src/main/java/com/jelastic/model/LogOut.java
@@ -1,31 +1,0 @@
-package com.jelastic.model;
-
-public class LogOut {
-    private int result;
-    private String error;
-    private Debug debug;
-
-    public int getResult() {
-        return result;
-    }
-
-    public void setResult(int result) {
-        this.result = result;
-    }
-
-    public String getError() {
-        return error;
-    }
-
-    public void setError(String error) {
-        this.error = error;
-    }
-
-    public Debug getDebug() {
-        return debug;
-    }
-
-    public void setDebug(Debug debug) {
-        this.debug = debug;
-    }
-}
