+++ src/main/java/com/jelastic/util/UploadResponse.java
@@ -1,0 +1,13 @@
+package com.jelastic.util;
+
+import lombok.Data;
+
+@Data
+public class UploadResponse {
+    private String name;
+    private String file;
+    private String error;
+
+    private int result;
+    private int size;
+}
