--- src/test/resources/authentication_error.json
@@ -1,1 +1,0 @@
-{"result":701,"error":"authentication failed","debug":{"time":249,"cpu":{"usage":1.0027938,"time":20}}}
