--- src/test/resources/authentication_ok.json
@@ -1,1 +1,0 @@
-{"uid":1163,"result":0,"session":"48bxaad71ccc7996325f3803311326b0247d","email":"Igor.Yova@gmail.com","name":"Igor Yova","debug":{"time":134,"cpu":{"usage":0.9312515,"time":10}}}
