--- src/test/resources/createobject_error.json
@@ -1,1 +1,0 @@
-{"result":3,"error":"invalid parameter [session]","debug":{"time":8,"cpu":{"usage":0.0,"time":0}}}
