--- src/test/resources/createobject_ok.json
@@ -1,1 +1,0 @@
-{"id":247,"result":0,"object":{"id":247,"developer":1163,"uploadDate":1307610094207},"debug":{"time":369,"cpu":{"usage":1.0157362,"time":30}}}
