--- src/test/resources/deploy_error.json
@@ -1,1 +1,0 @@
-{"response":null,"result":11,"error":"application [8129583aae37a4b556d36dbd56abbc68,8129583aae37a4b556d36dbd56abbc68] not exist","debug":{"time":479,"cpu":{"usage":0.0,"time":0}}}
