--- src/test/resources/uploader_error.json
@@ -1,1 +1,0 @@
-{"result":3,"error":"invalid param"}
