--- src/test/resources/uploader_ok.json
@@ -1,1 +1,0 @@
-{"result":0,"file":"http://api.hivext.com/xssu/rest/download/B2BTUQoDWkkwDDlERENeRFAZEEJdRzhMdCN0VkcTQ1IPAQBCY1VLQEg%3D","request":{"fid":"12345","appid":"8129583aae37a4b556d36dbd56abbc68","charset":"UTF-8"},"name":"jelastic-maven-plugin-1.0-SNAPSHOT.jar","size":14498}
